harel is the best
and bibi is the king

 good ver

+++ src/Beginner.java
@@ -93,7 +93,7 @@ public class Beginner implements State {
     }
 
     @Override
-    public void fileRequest(int fileSize) {
+    public void fileRequest(double fileSize) {
 
     }
