+++ src/Connected.java
@@ -78,7 +78,7 @@ public class Connected implements State{
     }
 
     @Override
-    public void fileRequest(int fileSize) {
+    public void fileRequest(double fileSize) {
 
     }
