+++ src/NetworkRegion.java
@@ -88,7 +88,7 @@ public class NetworkRegion implements State {
     }
 
     @Override
-    public void fileRequest(int fileSize) {
+    public void fileRequest(double fileSize) {
 
     }
