+++ src/Off.java
@@ -89,7 +89,7 @@ public class Off implements State {
     }
 
     @Override
-    public void fileRequest(int fileSize) {
+    public void fileRequest(double fileSize) {
 
     }
